+++ mdcx/controllers/main_window/save_config.py
@@ -282,6 +282,7 @@ def save_config(self: "MyMAinWindow"):
         (self.Ui.checkBox_nfo_criticrating, NfoInclude.CRITICRATING),
         (self.Ui.checkBox_nfo_actor, NfoInclude.ACTOR),
         (self.Ui.checkBox_nfo_all_actor, NfoInclude.ACTOR_ALL),
+        (self.Ui.checkBox_nfo_actor_tmdbid, NfoInclude.ACTOR_TMDBID),
         (self.Ui.checkBox_nfo_director, NfoInclude.DIRECTOR),
         (self.Ui.checkBox_nfo_series, NfoInclude.SERIES),
         (self.Ui.checkBox_nfo_tag, NfoInclude.TAG),
@@ -635,6 +636,8 @@ def save_config(self: "MyMAinWindow"):
     manager.config.fc2ppvdb = self.Ui.plainTextEdit_cookie_fc2ppvdb.toPlainText()  # fc2ppvdb cookie
     manager.config.javbus = self.Ui.plainTextEdit_cookie_javbus.toPlainText()  # javbus cookie
     manager.config.theporndb_api_token = self.Ui.lineEdit_api_token_theporndb.text()  # api token
+    manager.config.tmdb_api_base = self.Ui.lineEdit_tmdb_api_base.text().strip()  # TMDB API 地址
+    manager.config.tmdb_api_key = self.Ui.lineEdit_tmdb_api_key.text().strip()  # TMDB API Key
     if manager.config.javdb:
         manager.config.javdb = manager.config.javdb.replace("locale=en", "locale=zh")
     # endregion